backend/pyproject.toml
[project]
name = "aeterna-backend"
version = "0.1.0"
description = "FastAPI backend for the Echoes of Eternity AR experience."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "pydantic>=2.6.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-multipart>=0.0.9,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "ipython",
  "pytest",
  "pytest-asyncio"
]

[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

